feat: component rename
